move privateChats.ts and useVenueChat.ts to dedicated chat subfolder
